rename server test file so it is run
